+++ package.json
@@ -25,8 +25,14 @@
     "homepage": "https://github.com/nerdpruitt/nerdpruitt.github.io#readme",
     "devDependencies": {
         "@11ty/eleventy": "^0.12.1",
+        "css-loader": "^5.2.4",
+        "fibers": "^5.0.0",
+        "mini-css-extract-plugin": "^1.6.0",
         "npm-run-all": "^4.1.5",
         "rimraf": "^3.0.2",
+        "sass": "^1.32.13",
+        "sass-loader": "^11.1.1",
+        "style-loader": "^2.0.0",
         "webpack": "^5.36.0",
         "webpack-cli": "^4.6.0"
     }
