+++ src/_includes/headers/about-header.liquid
@@ -1,0 +1,4 @@
+<div class="page-header page-header--vertical">
+    <img class="page-header__img" srcset="/images/portrait.jpg 1x, /images/portrait@2x.jpg 2x" src="/images/portrait.jpg" alt="Portrait of Nick Pruitt">
+    <h1 class="page-header__title">{{ title }}</h1>
+</div>
