+++ src/_includes/layouts/base.liquid
@@ -1,0 +1,83 @@
+<!DOCTYPE html>
+<html lang="en" class="no-js">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    <title>Nick Pruitt</title>
+
+    <meta name="description" content="Nick Pruitt, front end developer, graphic designer, and your new friend.">
+
+    <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png">
+    <link rel="manifest" href="/site.webmanifest">
+    <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5">
+    <meta name="msapplication-TileColor" content="#da532c">
+    <meta name="theme-color" content="#ffffff">
+
+    <link rel="stylesheet" href="https://use.typekit.net/xyc1vcw.css">
+    <link rel="stylesheet" href="/dist/css/main.css">
+
+    <script>
+        document.documentElement.className = document.documentElement.className.replace(/\bno-js\b/g, '') + 'js';
+    </script>
+</head>
+<body>
+    <header class="site-header">
+        {% if pageHeader %}
+            {% include pageHeader %}
+        {% else %}
+            <div class="page-header page-header--bottom">
+                <h1 class="page-header__title">{{ pageTitle }}</h1>
+            </div>
+        {% endif %}
+
+        <div>
+            <nav class="nav-main" role="navigation">
+                <h1 class="nav-main__logo">nickpruitt</h1>
+                <ul class="nav-main__list">
+                    <li><a class="nav-main__link" href="{{ '/about' | url }}">about</a></li>
+                    <li><a class="nav-main__link" href="{{ '/portfolio' | url }}">portfolio</a></li>
+                    <li><a class="nav-main__link" href="{{ '/resume' | url }}">resume</a></li>
+                </ul>
+            </nav>
+        </div>
+    </header>
+
+    <main>
+        {{ content }}
+    </main>
+
+    <div class="site-footer-wrap">
+        <div class="site-footer__years">
+            <div class="site-footer__block" role="presentation"></div>
+            <footer class="site-footer">
+                <div class="site-footer__block" role="presentation"></div>
+
+                <small class="site-footer__small">
+                    2005&ndash;{% year %}
+                </small>
+
+                <div class="site-footer__block" role="presentation"></div>
+            </footer>
+            <div class="site-footer__block" role="presentation"></div>
+        </div>
+        <div class="site-footer__sarcastic-fine-print">
+            <div role="presentation"></div>
+            <small>
+                Made by Nick Pruitt. I would include a list of social media here, but aren’t we all better off without it?
+            </small>
+        </div>
+    </div>
+
+    <script src="/dist/js/main.js" async></script>
+
+    {% comment %} <script>
+        // Initialize the service worker
+        if (navigator && navigator.serviceWorker) {
+           navigator.serviceWorker.register('/sw.js');
+        }
+    </script> {% endcomment %}
+</body>
+</html>
