--- src/about.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" class="no-js">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-    <title>Nick Pruitt</title>
-
-    <meta name="description" content="Nick Pruitt, graphic designer, front end developer, and your new friend.">
-
-    <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png">
-    <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png">
-    <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png">
-    <link rel="manifest" href="/site.webmanifest">
-    <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5">
-    <meta name="msapplication-TileColor" content="#da532c">
-    <meta name="theme-color" content="#ffffff">
-
-    <link rel="stylesheet" href="/css/main.css">
-
-    <script>
-        document.documentElement.className = document.documentElement.className.replace(/\bno-js\b/g, '') + 'js';
-    </script>
-</head>
-<body>
-    <h1>I am a front end web developer and designer from Nashville, TN, and I like making things.</h1>
-
-    <p>For fifteen years, I have been designing and developing websites. I’ve been fortunate enough to do a little bit of everything over the course of my career. I have lead a team in graphic standards across a large e-commerce site, developed the front end of a travel agency booking system, created a mailing admin with dynamic template building, kept a legacy codebase up and running while rebuilding it piece by piece, created a patient managed care single page application, developed an HTML and CSS system that tightly integrated with a CMS powering over 600 web sites (before it was cool), and I have designed and developed a video game (just for the fun of it), to name a few highlights.</p>
-
-    <p>I have never neatly fit into a single category in our industry because I’m too dang curious. I was the kid who was always drawing and reading. In college I earned a BFA in Graphic Design and taught myself how to code while working a summer internship on the web team at a large corporation. I progressed from simple HTML and CSS to Sass to jQuery to BackboneJS and VueJS to writing most things from scratch these days.</p>
-
-    <p>I believe the fun in building software is in figuring things out, in learning something new, in crafting something elegant that I made myself and am proud of, and that I know will benefit and serve those who use it. I simply like solving problems and making things, and I plan to continue doing so for quite some time.</p>
-
-    <script src="/js/main.js" async></script>
-</body>
-</html>
