+++ src/about.liquid
@@ -1,0 +1,82 @@
+---
+layout: layouts/base.liquid
+title: About Me
+pageHeader: ../headers/about-header.liquid
+---
+
+<div class="grid-main">
+    <div class="grid-main__content">
+        <h2 class="page-title">I am a front end web developer and designer from Nashville, TN, and I like making things.</h2>
+    </div>
+</div>
+
+<section class="grid-main">
+    <div class="grid-main__gutter-l">
+        <h3 class="section-title">A Short History</h2>
+    </div>
+    <div class="grid-main__content">
+        <div class="columns u-margin--b-m">
+            <p class="dropcap">For fifteen years, I have been designing and developing websites. Over the course of my career, I’ve been fortunate enough to do a little bit of everything for a variety of companies large and small. I am now the grizzled old guy who has seen it all and speaks only in cryptic zen koans. I find I like this station of life quite a bit… I don’t even miss having hair!</p>
+
+            <p>Some highlights in my career include leading a team in graphic standards across a large e-commerce site, developing the front end of a travel agency-esque booking system, creating a mailing admin with dynamic template building and content creation, keeping a legacy codebase up and running while rebuilding it piece by piece, creating a patient managed care single page application, developing an HTML and CSS system that tightly integrated with a CMS powering over 600 web sites, and designing, developing, and creating all the art for a video game with a great friend of mine—just for the pure fun of it.</p>
+
+            <p>I have never neatly fit into a single category in our industry because I’m a very curious person. Once I become interested in something I have to learn all that I can about it. While I was earning a BFA in Graphic Design during my college days, I taught myself how to code while working a summer internship on the web team at a large corporation. I discovered that I enjoyed coding as much as I did design and art, and I never looked back. I started with simple HTML and CSS, venturing in the worlds of Sass and jQuery and various Content Management Systems as they hit the scene. I progressed into writing SPAs in BackboneJS and VueJS and discovered I really like writing JavaScript. These days I write most things from scratch in Vanilla JS because I think it’s more interesting and enjoyable.</p>
+        </div>
+
+        <p>Creative problem solving is why I find building software fun. I want to craft elegant things in which I can take pride, and that I know will benefit and serve the people who use it. My goal has never been to move up a corporate ladder or to manage people but rather to solve problems and to make things. This is where my true passion rests, and I plan to continue doing it for some time.</p>
+    </div>
+    <div class="grid-main__gutter-r"></div>
+</section>
+
+<hr>
+
+<section class="grid-main-indent">
+    <div class="grid-main__gutter-l">
+        <h3 class="section-title">Skills</h2>
+    </div>
+    <div class="grid-main__content">
+        <p>In the course of my career I have designed user flows for things like store checkouts and travel reservation systems. I have designed logos and worked with branding teams to design and implement marketing sites. I am also proficient in HTML and CSS dating back to handling notorious IE6 bugs (you can ask your parents about that one) and the revolution that responsive design brought to our industry. I am proficient in HTML, CSS, and JS, building scalable code that works across multiple apps and properties. I can pick up any JS framework you want, or write my own plugins from scratch. I love collaborating closely with backend developers to create features that delight the people who use them The web is amazing to me, and I love working with it. Here is a list of the types of things I enjoying creating.</p>
+
+        <ul class="columns">
+            <li>HTML</li>
+            <li>CSS/Sass</li>
+            <li>Vanilla JS</li>
+            <li>VueJS</li>
+            <li>Single Page Applications</li>
+            <li>Automated Testing (Cypress, Jest)</li>
+            <li>CMS Integration</li>
+            <li>Refactoring Legacy Systems</li>
+            <li>E-commerce Sites</li>
+            <li>Project Planning</li>
+            <li>Brand and Style Guide Design</li>
+            <li>Design Systems</li>
+            <li>UI Design</li>
+            <li>Prototyping</li>
+            <li>Marketing Sites</li>
+            <li>Dad jokes</li>
+            <li>Common sense</li>
+            <li>Being a sane, decent adult</li>
+        </ul>
+    </div>
+    <div class="grid-main__gutter-r"></div>
+</section>
+<section>
+        <h2>Colophon</h2>
+
+        <p>For a full breakdown of the design choices that make this site up, as well as the super cool tech running under the hood, check out the case study I wrote for this site.</p>
+
+        <h2>A Few of My Favorite Things</h2>
+
+        <p>Because life is about far more than what you do for a living, here is a list of some random things I enjoy. You should go enjoy things too!</p>
+
+        <ul>
+            <li>My three children</li>
+            <li>St. Thomas Aquinas &  G.K. Chesterton</li>
+            <li>Reading too many books at once</li>
+            <li>Old kung fu movies</li>
+            <li>Coffee, cocktails, and cigars</li>
+            <li>Japanese pro wrestling</li>
+            <li>Constant doodling</li>
+            <li>Vector artwork</li>
+        </ul>
+</section>
