+++ src/index.liquid
@@ -1,0 +1,5 @@
+---
+layout: layouts/base.liquid
+---
+
+<img src="/images/avatar.svg" alt="The angry yet lovely face of Nick Pruitt" width="256">
