+++ src/portfolio/index.liquid
@@ -1,0 +1,7 @@
+---
+layout: layouts/base.liquid
+---
+
+<h1>Portfolio</h1>
+
+<p><a href="{{ '/portfolio/capacity-managed-tickets' | url }}">Capacity Managed Tickets</a></p>
