+++ src/resume.liquid
@@ -1,0 +1,5 @@
+---
+layout: layouts/base.liquid
+---
+
+<h1>Resume</h1>
