+++ src/scss/base/_utility.scss
@@ -1,0 +1,5 @@
+@use '../functions' as _f;
+
+.u-margin--b-m {
+    margin-bottom: 2rem;
+}
