+++ src/scss/base/index.scss
@@ -1,0 +1,1 @@
+@forward 'utility';
