+++ src/scss/components/_footer.scss
@@ -1,0 +1,67 @@
+@use '../functions' as _f;
+
+.site-footer-wrap {
+    display: grid;
+    grid-gap: _f.rem(40);
+    grid-template-rows: auto auto;
+    padding-bottom: _f.rem(40);
+}
+
+.site-footer__years {
+    display: grid;
+    grid-template-columns: 1fr minmax(auto, var(--main-width)) 1fr;
+    grid-row: 1 / 2;
+}
+
+.site-footer {
+    display: grid;
+    grid-gap: _f.rem(40);
+    grid-template-columns: minmax(auto, _f.rem(407)) auto 1fr;
+}
+
+.site-footer__block {
+    position: relative;
+
+    &::before,
+    &::after {
+        content: '';
+        left: 0;
+        position: absolute;
+        right: 0;
+    }
+
+    &::before {
+        background-color: var(--yellow);
+        height: _f.rem(120);
+        bottom: _f.rem(72);
+    }
+
+    &::after {
+        background-color: var(--black);
+        height: _f.rem(32);
+        bottom: 0;
+    }
+}
+
+.site-footer__small {
+    align-self: end;
+    color: var(--red);
+    font-family: var(--font-family-heading);
+    font-size: _f.rem(64);
+    font-weight: var(--font-weight-black);
+    letter-spacing: var(--letter-spacing-caps);
+    line-height: var(--line-height-heading);
+    margin: 0;
+    text-transform: uppercase;
+    writing-mode: vertical-lr;
+}
+
+.site-footer__sarcastic-fine-print {
+    display: grid;
+    grid-gap: _f.rem(40);
+    grid-row: 2 / 3;
+    grid-template-columns: minmax(auto, _f.rem(407)) minmax(auto, _f.rem(500));
+    margin: 0 auto;
+    max-width: var(--main-width);
+    width: 100%;
+}
