+++ src/scss/components/_grid.scss
@@ -1,0 +1,157 @@
+@use '../functions' as _f;
+
+.grid-main {
+    display: grid;
+    grid-template-columns: 1fr minmax(65ch, 1040px) 1fr;
+}
+
+.grid-main-indent {
+    display: grid;
+    grid-template-columns: 250px minmax(65ch, 1040px) 1fr;
+}
+
+.grid-main__gutter-l {
+    grid-column: 1 / 2;
+    justify-self: end;
+}
+
+.grid-main__content {
+    grid-column: 2 / 3;
+}
+
+.grid-main__content--narrow {
+    max-width: _f.rem(520);
+}
+
+.grid-main__gutter-r {
+    grid-column: 3 / 4;
+}
+
+// https://betravis.github.io/shape-tools/path-to-polygon/
+.grid-img {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+}
+
+.grid-img__content {
+    &::before {
+        content: '';
+        float: right;
+        height: 408px;
+        shape-outside: polygon(80px 0px, 80px 408px, 0px 204px);
+        width: 80px;
+    }
+}
+
+.grid__img {
+    position: relative;
+    z-index: 1;
+
+    &::before {
+        background-image: url("data:image/svg+xml;charset=UTF-8,%3csvg height='408' viewBox='0 0 80 408' width='80' xmlns='http://www.w3.org/2000/svg'%3e%3cpath d='m80 0-80 203.501222 80 204.498778h-80v-408z' fill='%23fff' fill-rule='evenodd'/%3e%3c/svg%3e");
+        content: '';
+        height: 408px;
+        left: 0;
+        position: absolute;
+        top: 0;
+        width: 80px;
+        z-index: 2;
+    }
+
+    img {
+        height: 408px;
+        float: left;
+        margin-bottom: _f.rem(24);
+        width: auto;
+    }
+}
+
+.grid__img--left {
+    img {
+        float: right;
+    }
+
+    &::before {
+        background-image: url("data:image/svg+xml;charset=UTF-8,%3csvg height='408' viewBox='0 0 80 408' width='80' xmlns='http://www.w3.org/2000/svg'%3e%3cpath d='m80 0v408h-80l80-200-80-208z' fill='%23fff' fill-rule='evenodd'/%3e%3c/svg%3e");
+        left: auto;
+        right: 0;
+    }
+
+    .grid-img__caption {
+        margin-left: 0;
+        margin-right: 80px;
+    }
+}
+
+.grid-img__content--right {
+    &::before {
+        float: left;
+        shape-outside: polygon(0px 0px, 80px 208px, 0px 408px);
+    }
+}
+
+.grid-img__caption {
+    clear: both;
+    color: var(--caption-color);
+    font-size: _f.rem(18);
+    font-weight: var(--font-weight-light);
+    margin-left: 80px;
+}
+
+.grid-img-full {
+    margin: 0 auto;
+    max-width: var(--main-width);
+}
+
+.page-title {
+    font-family: var(--font-family-heading);
+    font-size: _f.rem(40);
+    font-weight: var(--font-weight-bold);
+    line-height: var(--line-height-heading);
+    margin-top: 0;
+    max-width: _f.rem(868);
+}
+
+.section-title {
+    color: var(--red);
+    font-family: var(--font-family-heading);
+    font-size: _f.rem(40);
+    font-weight: var(--font-weight-black);
+    letter-spacing: var(--letter-spacing-caps);
+    line-height: var(--line-height-heading);
+    margin: 0.5em;
+    text-transform: uppercase;
+    writing-mode: vertical-lr;
+}
+
+.previous,
+.next {
+    color: var(--red);
+    font-family: var(--font-family-heading);
+    font-size: _f.rem(40);
+    display: inline-flex;
+    letter-spacing: var(--letter-spacing-caps);
+    margin-top: -32px;
+    position: fixed;
+    text-decoration: none;
+    text-transform: uppercase;
+    top: 50%;
+
+    span {
+        opacity: 0;
+        transition: opacity 0.25s ease-in-out;
+    }
+
+    &:hover span {
+        opacity: 1;
+    }
+}
+
+.previous {
+    flex-direction: row-reverse;
+    left: 2rem;
+}
+
+.next {
+    right: 2rem;
+}
