+++ src/scss/components/_header.scss
@@ -1,0 +1,98 @@
+@use '../functions' as _f;
+
+.site-header {
+    display: grid;
+    grid-template-columns: 1fr _f.rem(292);
+    margin: 0 auto 2rem;
+    max-width: var(--main-width);
+}
+
+// ---------------------
+// Main site navigation
+// ---------------------
+.nav-main {
+    background-color: var(--green);
+    padding: _f.rem(32) _f.rem(40);
+}
+
+.nav-main__logo {
+    color: var(--white);
+    font-size: _f.rem(36);
+    line-height: 1;
+}
+
+.nav-main__list {
+    list-style: none;
+    margin: 0;
+    padding: 0;
+
+    > li {
+        margin-bottom: _f.rem(16);
+
+        &:last-child {
+            margin-bottom: 0;
+        }
+    }
+}
+
+.nav-main__link {
+    color: var(--white);
+    font-size: _f.rem(24);
+    position: relative;
+    text-decoration: none;
+    z-index: 5;
+
+    &::after {
+        background-color: var(--black);
+        bottom: 2px;
+        content: '';
+        left: -2px;
+        height: 4px;
+        position: absolute;
+        right: -2px;
+        z-index: -1;
+    }
+}
+
+
+// ---------------------
+// Page Header
+// ---------------------
+
+.page-header {
+    display: grid;
+    grid-column-gap: _f.rem(40);
+    grid-template-columns: auto 1fr;
+    padding-top: _f.rem(80);
+}
+
+.page-header__img {
+    height: auto;
+    max-width: 407px;
+    width: 100%;
+}
+
+.page-header--vertical {
+    .page-header__title {
+        align-self: end;
+        color: var(--red);
+        font-family: var(--font-family-heading);
+        font-size: _f.rem(64);
+        font-weight: var(--font-weight-black);
+        letter-spacing: var(--letter-spacing-caps);
+        line-height: var(--line-height-heading);
+        margin: 0;
+        text-transform: uppercase;
+        writing-mode: vertical-lr;
+    }
+}
+
+.page-header--bottom {
+    .page-header__title {
+        align-self: end;
+        font-size: _f.rem(54);
+        font-weight: var(--font-weight-black);
+        line-height: var(--line-height-heading);
+        margin: 0;
+    }
+}
