+++ src/scss/components/_hr.scss
@@ -1,0 +1,7 @@
+hr {
+    background: var(--black);
+    border: 0;
+    height: 1rem;
+    max-width: var(--main-width);
+    width: 100%;
+}
