+++ src/scss/components/index.scss
@@ -1,0 +1,4 @@
+@forward 'footer';
+@forward 'grid';
+@forward 'header';
+@forward 'hr';
