+++ src/scss/config/_colors.scss
@@ -1,0 +1,11 @@
+:root {
+    --black: #21231A;
+    --gray: #AFB0AA;
+    --green: #7A8040;
+    --red: #E13C38;
+    --white: #FFF;
+    --yellow: #F8EC49;
+
+    --main-bg-color: var(--white);
+    --caption-color: var(--gray);
+}
