+++ src/scss/config/_layout.scss
@@ -1,0 +1,3 @@
+:root {
+    --grid-column-gap: 2vw;
+}
