+++ src/scss/config/_spacing.scss
@@ -1,0 +1,5 @@
+@use '../functions' as _f;
+
+:root {
+    --main-width: #{_f.rem(1200)};
+}
