+++ src/scss/config/_typography.scss
@@ -1,0 +1,14 @@
+:root {
+    // Font family
+    --font-family-dropcap: interstate-condensed, sans-serif;
+    --font-family-heading: interstate-condensed, sans-serif;
+    --font-family-main: interstate, sans-serif;
+    // Font weights
+    --font-weight-light: 400;
+    --font-weight-regular: 500;
+    --font-weight-bold: 700;
+    --font-weight-black: 800;
+
+    --line-height-heading: 1;
+    --letter-spacing-caps: 0.025em;
+}
