+++ src/scss/config/index.scss
@@ -1,0 +1,4 @@
+@forward 'colors';
+@forward 'layout';
+@forward 'spacing';
+@forward 'typography';
