+++ src/scss/functions/index.scss
@@ -1,0 +1,1 @@
+@forward 'rem';
