+++ src/scss/index.scss
@@ -1,0 +1,102 @@
+@forward './config';
+@forward './base';
+@forward './components';
+
+@use './functions' as _f;
+
+*,
+*::before,
+*::after {
+    box-sizing: border-box;
+}
+
+body,
+html,
+figure {
+    margin: 0;
+    padding: 0;
+}
+
+body,
+html {
+    min-height: calc(100vh - 0.5rem);
+}
+
+html {
+    background-color: var(--black);
+}
+
+body {
+    background-color: var(--main-bg-color);
+    border-radius: 0.25rem;
+    color: var(--black);
+    font-family: var(--font-family-main);
+    font-size: _f.rem(18);
+    line-height: 1.5;
+    margin: 0.25rem;
+}
+
+img {
+    display: block;
+    margin: 0 auto;
+}
+
+h1 {
+    margin-bottom: 1rem;
+    margin-top: 0;
+}
+
+/* Sizzle */
+.js .fade-in {
+    opacity: 1;
+    transition: opacity 0.3s ease-in-out;
+}
+
+.js .is-hidden {
+    opacity: 0;
+}
+
+
+// main {
+//     padding: 2rem;
+// }
+
+p {
+    margin-bottom: 0;
+    margin-top: 0;
+}
+
+p + p {
+    text-indent: 1.5em;
+}
+
+@media (max-width: 528px) {
+    h1 {
+        --minFontSize: 1.5rem;
+        --maxFontSize: 3rem;
+        --scaler: 10vw;
+        font-size: clamp( var(--minFontSize), var(--scaler), var(--maxFontSize) );
+    }
+
+    p {
+        font-size: 1rem;
+    }
+
+    // main {
+    //     padding: 1rem;
+    // }
+}
+
+.dropcap::first-letter {
+    float: left;
+    font-family: var(--font-family-dropcap);
+    font-size: 3.25rem;
+    line-height: 1;
+    padding-right: 8px;
+}
+
+.columns {
+    column-count: 2;
+    column-gap: var(--grid-column-gap);
+    // column-rule: var(--border-width-thin) solid var(--color-border);
+}
